requirements.txt
streamlit
sentence-transformers
faiss-cpu
python-dotenv
requests
